Added a method to retrieve a key from the Keychain for an identifier
I have a use case where I want to store an API token in the keychain after a user has authenticated. This API token is then used for requests from the app to the API. I wanted to be able to query the keychain for the API token identifier and retrieve the token. If the token does not exist then prompt the user to login again to get a new token.